Put BeliefVector into it's own file
